+++ python/cusignal/convolution/convolve.py
@@ -16,15 +16,29 @@ from cupyx.scipy import fftpack
 import sys
 
 from .. import _signaltools
-from ..utils.fftpack_helper import _init_nd_shape_and_axes_sorted, \
-    next_fast_len
-from .convolution_utils import _inputs_swap_needed, _numeric_arrays, \
-    _centered, _fftconv_faster, _timeit_fast
+from ..utils.fftpack_helper import (
+    _init_nd_shape_and_axes_sorted,
+    next_fast_len,
+)
+from .convolution_utils import (
+    _inputs_swap_needed,
+    _numeric_arrays,
+    _centered,
+    _fftconv_faster,
+    _timeit_fast,
+)
 
 _modedict = {"valid": 0, "same": 1, "full": 2}
 
 
-def convolve(in1, in2, mode="full", method="auto"):
+def convolve(
+    in1,
+    in2,
+    mode="full",
+    method="auto",
+    cp_stream=cp.cuda.stream.Stream(null=True),
+    autosync=True,
+):
     """
     Convolve two N-dimensional arrays.
 
@@ -62,6 +76,17 @@ def convolve(in1, in2, mode="full", method="auto"):
         ``auto``
            Automatically chooses direct or Fourier method based on an estimate
            of which is faster (default).
+    cp_stream : CuPy stream, optional
+        Option allows upfirdn to run in a non-default stream. The use
+        of multiple non-default streams allow multiple kernels to
+        run concurrently. Default is cp.cuda.stream.Stream(null=True)
+        or default stream.
+    autosync : bool, optional
+        Option to automatically synchronize cp_stream. This will block
+        the host code until kernel is finished on the GPU. Setting to
+        false will allow asynchronous operation but might required
+        manual synchronize later `cp_stream.synchronize()`.
+        Default is True.
 
     Returns
     -------
@@ -140,7 +165,7 @@ def convolve(in1, in2, mode="full", method="auto"):
             volume, kernel = kernel, volume
 
         return _signaltools._convolve(
-            volume, kernel, True, swapped_inputs, mode
+            volume, kernel, True, swapped_inputs, mode, cp_stream, autosync
         )
 
     else:
@@ -185,7 +210,6 @@ def fftconvolve(in1, in2, mode="full", axes=None):
     axes : int or array_like of ints or None, optional
         Axes over which to compute the convolution.
         The default is over all axes.
-
 
     Returns
     -------
@@ -323,6 +347,7 @@ def convolve2d(
     boundary="fill",
     fillvalue=0,
     cp_stream=cp.cuda.stream.Stream(null=True),
+    autosync=True,
     use_numba=False,
 ):
     """
@@ -357,6 +382,21 @@ def convolve2d(
            symmetrical boundary conditions.
     fillvalue : scalar, optional
         Value to fill pad input arrays with. Default is 0.
+    cp_stream : CuPy stream, optional
+        Option allows upfirdn to run in a non-default stream. The use
+        of multiple non-default streams allow multiple kernels to
+        run concurrently. Default is cp.cuda.stream.Stream(null=True)
+        or default stream.
+    autosync : bool, optional
+        Option to automatically synchronize cp_stream. This will block
+        the host code until kernel is finished on the GPU. Setting to
+        false will allow asynchronous operation but might required
+        manual synchronize later `cp_stream.synchronize()`.
+        Default is True.
+    use_numba : bool, optional
+        Option to use Numba CUDA kernel or raw CuPy kernel. Raw CuPy
+        can yield performance gains over Numba. Default is False.
+
     Returns
     -------
     out : ndarray
@@ -400,14 +440,7 @@ def convolve2d(
         in1, in2 = in2, in1
 
     out = _signaltools._convolve2d(
-        in1,
-        in2,
-        1,
-        mode,
-        boundary,
-        fillvalue,
-        cp_stream=cp_stream,
-        use_numba=use_numba,
+        in1, in2, 1, mode, boundary, fillvalue, cp_stream, autosync, use_numba,
     )
     return out