+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - PR #66 - Add CuPy Module for 2d upfirdn, along with tests/benchmarks
 - PR #73 - Local gpuCI build script
 - PR #75 - Add accelerated `lfilter` method.
+- PR #82 - Implement `autosync` to synchronize raw kernels by default
 
 ## Improvements
 - PR #40 - Ability to specify time/freq domain for resample.